% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Title}
\usage{
extract_data(
  data,
  years = NULL,
  ages = "All Ages",
  metrics = "Number",
  measures = "Incidence",
  gender = "Both",
  rounds = 0,
  EAPC = F,
  EC_change = F
)
}
\arguments{
\item{data}{data}

\item{years}{select the year to analysis}

\item{ages}{such as 'Age-standardized'}

\item{metrics}{such as 'Incidence'}

\item{measures}{look up the website}

\item{gender}{look up the website}

\item{rounds}{rounds}

\item{EAPC}{calculate EAPC}

\item{EC_change}{CALCULATE EC_change}
}
\value{
the data
}
\description{
Title
}
\examples{
EC;extract_data(EC,1990)
}
